+++ resource-registry/CHANGELOG.md
@@ -1,9 +1,20 @@
-# Revision history of strict-checked-vars
+# Revision history of `resource-registry`
 
 ## Next version
 
 * Define `transferRegistry` for moving all resources from one registry to a
   different one.
+
+* Cancel registered threads in the registry before closing the registry. This
+  prevents race conditions where the registry gets closed but the running async
+  thread still tries to allocate something in the registry before being
+  cancelled, which would result in an exception.
+
+* Expose `allocateThread` to register another thread in a registry, such that it
+  is cancelled before closing the registry, even if it belongs to a separate
+  registry.
+
+* Add the label of the thread to `Context`.
 
 ## 0.1.1.0 — 2025-05-15